+++ lib/s_extensions.dart
@@ -8,6 +8,7 @@ export 'extensions/datetime_ext.dart';
 export 'extensions/list_ext.dart';
 export 'extensions/map_ext.dart';
 export 'extensions/number_ext.dart';
+export 'extensions/object_ext.dart';
 export 'extensions/string_ext.dart';
 export 'extensions/widget_ext.dart';